Fix compiler warning on SunOS
pcscdaemon.c: In function 'main':
pcscdaemon.c:402:5: warning: format '%d' expects argument of type 'int', but argument 6 has type 'pid_t' [-Wformat=]
     Log2(PCSC_LOG_CRITICAL,
     ^

+++ src/pcscdaemon.c
@@ -491,7 +491,7 @@ int main(int argc, char **argv)
 				Log1(PCSC_LOG_CRITICAL,
 					"file " PCSCLITE_CSOCK_NAME " already exists.");
 				Log2(PCSC_LOG_CRITICAL,
-					"Another pcscd (pid: %d) seems to be running.", pid);
+					"Another pcscd (pid: %ld) seems to be running.", (long)pid);
 				return EXIT_FAILURE;
 			}
 			else
